Add validation for bookId and commentId into CommentRestfulAPI class

+++ assignment4/src/main/java/com/ngoc/bookmanagement/restfulAPI/CommentRestfulAPI.java
@@ -1,7 +1,8 @@
-package com.ngoc.bookmanagement.restfulApi;
+package com.ngoc.bookmanagement.restfulAPI;
 
 import com.ngoc.bookmanagement.model.Comment;
 import com.ngoc.bookmanagement.model.Message;
+import com.ngoc.bookmanagement.repository.BookRepository;
 import com.ngoc.bookmanagement.repository.CommentRepository;
 import com.ngoc.bookmanagement.validation.GroupCommentCreate;
 import com.ngoc.bookmanagement.validation.GroupCommentUpdate;
@@ -22,6 +23,9 @@ public class CommentRestfulAPI {
     @Autowired
     private CommentRepository commentRepository;
 
+    @Autowired
+    private BookRepository bookRepository;
+
     private static final Validator validator;
 
     static {
@@ -31,17 +35,30 @@ public class CommentRestfulAPI {
         factory.close();
     }
 
-    // API get a comment by id
-    @GetMapping(value = "/api/comment/{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<?> getComment(@PathVariable("id") long id){
-        Comment commentIsSelected = commentRepository.findById(id).get();
+    // API get a comment by commentId
+    @GetMapping(value = "/api/comment/{commentId}", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<?> getComment(@PathVariable("commentId") long commentId){
 
+        Message message = new Message();
+        if(!commentRepository.existsById(commentId)){
+            message.getContent().put("message", "Comment isn't exist");
+            return new ResponseEntity<>(message.getContent(), HttpStatus.NOT_FOUND);
+        }
+
+        Comment commentIsSelected = commentRepository.getCommentById(commentId);
         return new ResponseEntity<>(commentIsSelected, HttpStatus.OK);
     }
 
     // API get all comments of book by bookId
     @GetMapping(value = "/api/book/{bookId}/comment", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<?> getAllCommentOfBook(@PathVariable("bookId") long bookId){
+
+        Message message = new Message();
+        if(!bookRepository.existsById(bookId)){
+            message.getContent().put("message", "Book isn't exist");
+            return new ResponseEntity<>(message.getContent(), HttpStatus.NOT_FOUND);
+        }
+
         List<Comment> commentList = commentRepository.getAllByBookId(bookId);
 
         return new ResponseEntity<>(commentList, HttpStatus.OK);
@@ -50,7 +67,14 @@ public class CommentRestfulAPI {
     // API create new comment for book, which rely on bookId
     @PostMapping(value = "/api/book/{bookId}/comment", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<?> createComment(@PathVariable("bookId") long bookId, @RequestBody Comment commentParam){
+
         Message message = new Message();
+
+        if(!bookRepository.existsById(bookId)){
+            message.getContent().put("message", "Book isn't exist");
+            return new ResponseEntity<>(message.getContent(), HttpStatus.NOT_FOUND);
+        }
+
         Set<ConstraintViolation<Comment>> constraintViolations = validator.validate(commentParam, GroupCommentCreate.class);
 
         if(constraintViolations.size() > 0){
@@ -73,6 +97,12 @@ public class CommentRestfulAPI {
     public ResponseEntity<?> updateComment(@PathVariable("commentId") long commentId,
                                            @RequestBody Comment commentParam){
         Message message = new Message();
+
+        if(!commentRepository.existsById(commentId)){
+            message.getContent().put("message", "Comment isn't exist");
+            return new ResponseEntity<>(message.getContent(), HttpStatus.NO_CONTENT);
+        }
+
         Set<ConstraintViolation<Comment>> constraintViolations = validator.validate(commentParam, GroupCommentUpdate.class);
 
         if(constraintViolations.size() > 0){
@@ -86,7 +116,6 @@ public class CommentRestfulAPI {
         Comment commentIsSelected = commentRepository.findById(commentId).get();
         commentIsSelected.setUpdatedAt(new Date());
         commentIsSelected.setMessage(commentParam.getMessage());
-
         commentRepository.save(commentIsSelected);
 
         return new ResponseEntity<>(commentIsSelected, HttpStatus.OK);
@@ -95,11 +124,16 @@ public class CommentRestfulAPI {
     // API delete a comment, which rely on commentId
     @DeleteMapping(value = "/api/comment/{commentId}", produces = {MediaType.APPLICATION_JSON_VALUE})
     public ResponseEntity<?> deleteComment(@PathVariable("commentId") long commentId){
-        commentRepository.deleteById(commentId);
-
         Message message = new Message();
+
+        if(!commentRepository.existsById(commentId)){
+            message.getContent().put("message", "Comment isn't exist");
+            return new ResponseEntity<>(message.getContent(), HttpStatus.NOT_FOUND);
+        }
+
+        commentRepository.deleteById(commentId);
         message.getContent().put("message", "Delete comment successfully");
 
-        return new ResponseEntity<>(message, HttpStatus.OK);
+        return new ResponseEntity<>(message.getContent(), HttpStatus.OK);
     }
 }
